mengubah int menjadi float

+++ pertemuan 4.cpp
@@ -24,7 +24,7 @@ int perkalian( int a, int b)
 {
     return a * b;
 }
-int pembagian( int a, int b)
+float pembagian( int a, int b)
 { 
     return a/b;
 }
